chore: rename app path `studio` to `sva-studio-react`
- renamed the Nx app project and folder via `nx move` and updated app metadata (apps/sva-studio-react/project.json:1, apps/sva-studio-react/package.json:1, apps/sva-studio-react/.cta.json:1)
- updated docs/specs and guidance to reference sva-studio-react (AGENTS.md:15, openspec/project.md:33, openspec/specs/monorepo-structure/spec.md:18, docs/monorepo.md:4)
- updated the lockfile importer path (pnpm-lock.yaml:30)
- ran required lint/type/unit tests
