+++ apps/sva-studio-react/.cta.json
@@ -1,5 +1,5 @@
 {
-  "projectName": "studio",
+  "projectName": "sva-studio-react",
   "mode": "code-router",
   "typescript": true,
   "tailwind": true,
@@ -9,8 +9,5 @@
   "addOnOptions": {},
   "version": 1,
   "framework": "react-cra",
-  "chosenAddOns": [
-    "start",
-    "nitro"
-  ]
+  "chosenAddOns": ["start", "nitro"]
 }
