+++ apps/sva-studio-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "studio",
+  "name": "sva-studio-react",
   "private": true,
   "type": "module",
   "scripts": {
