+++ apps/sva-studio-react/public/manifest.json
