+++ apps/sva-studio-react/src/routes/index.tsx
@@ -1,4 +1,4 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from '@tanstack/react-router';
 
 export const HomePage = () => {
   return (
@@ -9,23 +9,14 @@ export const HomePage = () => {
           className="h-16 w-16 rounded-full border border-slate-800 bg-slate-900 p-2"
           alt="SVA Studio"
         />
-        <h1 className="text-4xl font-semibold tracking-tight sm:text-5xl">
-          SVA Studio
-        </h1>
+        <h1 className="text-4xl font-semibold tracking-tight sm:text-5xl">SVA Studio</h1>
         <p className="max-w-2xl text-lg text-slate-300">
-          Smart Village App Self-Service Plattform für Inhalte, Module und
-          Erweiterungen.
+          Smart Village App Self-Service Plattform für Inhalte, Module und Erweiterungen.
         </p>
         <div className="flex flex-wrap items-center justify-center gap-3 text-sm text-slate-400">
-          <span className="rounded-full border border-slate-800 px-4 py-2">
-            Nx Monorepo
-          </span>
-          <span className="rounded-full border border-slate-800 px-4 py-2">
-            TanStack Start
-          </span>
-          <span className="rounded-full border border-slate-800 px-4 py-2">
-            Core + Plugins
-          </span>
+          <span className="rounded-full border border-slate-800 px-4 py-2">Nx Monorepo</span>
+          <span className="rounded-full border border-slate-800 px-4 py-2">TanStack Start</span>
+          <span className="rounded-full border border-slate-800 px-4 py-2">Core + Plugins</span>
         </div>
         <a
           className="rounded-full bg-emerald-400 px-6 py-3 text-sm font-semibold text-slate-900 transition hover:bg-emerald-300"
@@ -37,7 +28,7 @@ export const HomePage = () => {
         </a>
       </header>
     </div>
-  )
-}
+  );
+};
 
-export const Route = createFileRoute('/')({ component: HomePage })
+export const Route = createFileRoute('/')({ component: HomePage });
