+++ apps/sva-studio-react/tailwind.config.cjs
@@ -1,9 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: [
-    './src/**/*.{ts,tsx}',
-    '../../packages/**/*.{ts,tsx}',
-  ],
+  content: ['./src/**/*.{ts,tsx}', '../../packages/**/*.{ts,tsx}'],
   theme: {
     extend: {},
   },
